fix: validasi makan, tapi belum validasi tipe inputannya sesuai

+++ src/Memento/WalikotaMemento.hpp
@@ -12,7 +12,7 @@ using namespace std;
 
 class WalikotaMemento : public Memento{
     private:
-        map<string, float> map_pemain_gulden;
+        map<string, int> map_pemain_gulden;
         Pemain* created_pemain;
 
     public:
